Moved fixtures and main test functions to separate scripts
