+++ tests/test_filesys.py
@@ -7,6 +7,8 @@ import pickle
 
 from xlab.filesys import find_root_dir, Directories, MetadataLoader, HashmapLoader
 
+from .fixtures import project_setup
+
 
 
 ##### FIXTURES
@@ -44,34 +46,6 @@ def dir_structure(tmpdir):
     b_deep = b_a.mkdir('a').mkdir('a').mkdir('a')
 
     return tmpdir
-
-
-@pytest.fixture()
-def project_setup(tmpdir, request):
-    subdirs = request.param
-
-    os.makedirs(os.path.join(tmpdir, '.exp'), exist_ok=True)
-
-    path = tmpdir
-    for subdir in subdirs:
-        relative_path = os.path.join(*subdir) if type(subdir) == list else subdir
-        path = os.path.join(tmpdir, relative_path)
-
-        os.makedirs(path, exist_ok=True)
-    
-    future_curdir = path
-
-    dirname = os.path.dirname(os.path.realpath(__file__))
-    template_path = os.path.join(dirname, 'script_templates', 'main.py')
-    script_path = os.path.join(future_curdir, 'main.py')
-
-    shutil.copyfile(template_path, script_path)
-
-    return {
-        'root': tmpdir,
-        'curdir': future_curdir,
-        'script': script_path,
-    }
